Create filler.py
The class contains functions designed to fill nans and outliers, which can be determined by the interquartile range or by the 3-sigma rule.

It also contains the author's method for filling NaN and outliers - RDBR (Random Distribution By Ratio). The bottom line is this: the percentage of each normal value (not NaN or outliers) will be the probability of filling in that value instead of a specific NaN or outlier. In theory, this should help evenly fill in any distribution.